+++ plugins/cost-insights/src/types/ChangeStatistic.ts
@@ -14,52 +14,9 @@
  * limitations under the License.
  */
 
-import { Cost } from './Cost';
-import { MetricData } from './MetricData';
-import { aggregationSort } from '../utils/sort';
-
 export interface ChangeStatistic {
   // The ratio of change from one duration to another, expressed as: (newSum - oldSum) / oldSum
   ratio: number;
   // The actual USD change between time periods (can be negative if costs decreased)
   amount: number;
-}
-
-export const EngineerThreshold = 0.5;
-
-export enum ChangeThreshold {
-  upper = 0.05,
-  lower = -0.05,
-}
-
-export enum Growth {
-  Negligible,
-  Savings,
-  Excess,
-}
-
-// Used by <CostGrowth /> for displaying status colors
-export function growthOf(amount: number, ratio: number) {
-  if (amount >= EngineerThreshold && ratio >= ChangeThreshold.upper) {
-    return Growth.Excess;
-  }
-
-  if (amount >= EngineerThreshold && ratio <= ChangeThreshold.lower) {
-    return Growth.Savings;
-  }
-
-  return Growth.Negligible;
-}
-
-// Used by <CostOverviewCard /> for displaying engineer totals
-export function getComparedChange(
-  dailyCost: Cost,
-  metricData: MetricData,
-): ChangeStatistic {
-  const ratio = dailyCost.change.ratio - metricData.change.ratio;
-  const amount = dailyCost.aggregation.slice().sort(aggregationSort)[0].amount;
-  return {
-    ratio: ratio,
-    amount: amount * ratio,
-  };
 }
